# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Athlentry API", "version": "1.0.0" },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/v1/orgs/sport-templates": {
      "get": {
        "operationId": "get_api_v1_orgs_sport_templates",
        "summary": "List built-in sport templates",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "key": { "type": "string" },
                      "name": { "type": "string" }
                    },
                    "required": ["key", "name"],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/api/v1/orgs/{orgId}/credential-types": {
      "get": {
        "operationId": "get_api_v1_orgs_orgId_credential_types",
        "summary": "List organization safety requirements",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "key": { "type": "string" },
                      "name": { "type": "string" },
                      "verification": {
                        "type": "string",
                        "enum": [
                          "document_upload",
                          "attestation",
                          "provider",
                          "manual_staff"
                        ]
                      },
                      "validityMonths": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 120
                      },
                      "blocksActivation": { "type": "boolean" },
                      "active": { "type": "boolean" },
                      "version": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "id",
                      "key",
                      "name",
                      "verification",
                      "validityMonths",
                      "blocksActivation",
                      "active",
                      "version"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        },
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      }
    },
    "/api/v1/orgs/{orgId}/credential-types/{credentialId}": {
      "patch": {
        "operationId": "patch_api_v1_orgs_orgId_credential_types_credentialId",
        "summary": "Update organization safety requirement",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "key": { "type": "string" },
                    "name": { "type": "string" },
                    "verification": {
                      "type": "string",
                      "enum": [
                        "document_upload",
                        "attestation",
                        "provider",
                        "manual_staff"
                      ]
                    },
                    "validityMonths": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 120
                    },
                    "blocksActivation": { "type": "boolean" },
                    "active": { "type": "boolean" },
                    "version": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "name",
                    "verification",
                    "validityMonths",
                    "blocksActivation",
                    "active",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        },
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "validityMonths": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 120
                  },
                  "blocksActivation": { "type": "boolean" },
                  "active": { "type": "boolean" },
                  "version": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "name",
                  "validityMonths",
                  "blocksActivation",
                  "active",
                  "version"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{orgId}/members/{memberId}/roles": {
      "patch": {
        "operationId": "patch_api_v1_orgs_orgId_members_memberId_roles",
        "summary": "Update organization member roles",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "owner",
                          "admin",
                          "registrar",
                          "finance",
                          "scheduler",
                          "compliance",
                          "communications",
                          "director",
                          "evaluator",
                          "volunteer_coordinator",
                          "reporter"
                        ]
                      }
                    },
                    "pendingMfa": { "type": "boolean" },
                    "version": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": ["accountId", "roles", "pendingMfa", "version"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        },
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "roles": {
                    "minItems": 1,
                    "maxItems": 11,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "registrar",
                        "finance",
                        "scheduler",
                        "compliance",
                        "communications",
                        "director",
                        "evaluator",
                        "volunteer_coordinator",
                        "reporter"
                      ]
                    }
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": ["roles", "expectedVersion"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/slug-availability": {
      "get": {
        "operationId": "get_api_v1_orgs_slug_availability",
        "summary": "Check organization slug",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 40,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                    },
                    "available": { "type": "boolean" }
                  },
                  "required": ["slug", "available"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 3,
              "maxLength": 40,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ]
      }
    },
    "/api/v1/orgs": {
      "post": {
        "operationId": "post_api_v1_orgs",
        "summary": "Create organization",
        "tags": ["orgs"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "slug": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 40,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                    },
                    "status": { "type": "string", "const": "onboarding" }
                  },
                  "required": ["id", "slug", "status"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRule" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 160
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 40,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "club",
                      "league",
                      "association",
                      "academy",
                      "school",
                      "parks_rec",
                      "tournament_operator",
                      "other"
                    ]
                  },
                  "timezone": { "type": "string" },
                  "address": {
                    "type": "object",
                    "properties": {
                      "line1": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 160
                      },
                      "line2": { "type": "string", "maxLength": 160 },
                      "city": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 100
                      },
                      "region": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2,
                        "pattern": "^[A-Z]{2}$"
                      },
                      "postalCode": {
                        "type": "string",
                        "pattern": "^\\d{5}(?:-\\d{4})?$"
                      },
                      "country": { "type": "string", "const": "US" }
                    },
                    "required": [
                      "line1",
                      "city",
                      "region",
                      "postalCode",
                      "country"
                    ],
                    "additionalProperties": false
                  },
                  "sportKeys": {
                    "minItems": 1,
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    }
                  }
                },
                "required": [
                  "name",
                  "slug",
                  "kind",
                  "timezone",
                  "address",
                  "sportKeys"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-athlentry_session"
      },
      "bearerAuth": { "type": "http", "scheme": "bearer" }
    },
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNDER_13",
                  "WEAK_PASSWORD",
                  "INVALID_CREDENTIALS",
                  "VERIFICATION_REQUIRED",
                  "INVALID_TOKEN",
                  "NOT_FOUND",
                  "UNAUTHENTICATED",
                  "REAUTH_REQUIRED",
                  "FORBIDDEN",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "DEPENDENCY_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": { "type": "string" },
              "fields": {
                "type": "object",
                "propertyNames": { "type": "string" },
                "additionalProperties": { "type": "string" }
              }
            },
            "required": ["code", "message"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "Unauthenticated": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict or stale version",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "BusinessRule": {
        "description": "Business rule rejected request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "Unavailable": {
        "description": "Dependency unavailable",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      }
    }
  }
}